Harmonized the readers for the SINTED Ocean test data export format based on Matlab .mat files versions before and after 7.3.
